requirements.txt
customtkinter
cryptography
Pillow
pyperclip
